LfTopoConfig moved from impl
