+++ src/main/java/com/powsybl/openloadflow/network/LfNetwork.java
@@ -12,7 +12,6 @@ import com.google.common.base.Stopwatch;
 import com.powsybl.commons.reporter.Reporter;
 import com.powsybl.openloadflow.graph.GraphConnectivity;
 import com.powsybl.openloadflow.graph.GraphConnectivityFactory;
-import com.powsybl.openloadflow.network.impl.LfTopoConfig;
 import com.powsybl.openloadflow.util.PerUnit;
 import com.powsybl.openloadflow.util.Reports;
 import org.anarres.graphviz.builder.GraphVizGraph;
