+++ src/main/java/com/powsybl/openloadflow/network/LfTopoConfig.java
@@ -4,7 +4,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.openloadflow.network.impl;
+package com.powsybl.openloadflow.network;
 
 import com.powsybl.iidm.network.Switch;
 
@@ -50,11 +50,11 @@ public class LfTopoConfig {
         return !(switchesToOpen.isEmpty() && switchesToClose.isEmpty() && busIdsToLose.isEmpty());
     }
 
-    public boolean retainPtc(String branchId) {
+    public boolean isRetainedPtc(String branchId) {
         return branchIdsWithPtcToRetain.contains(branchId);
     }
 
-    public boolean retainRtc(String branchId) {
+    public boolean isRetainedRtc(String branchId) {
         return branchIdsWithRtcToRetain.contains(branchId);
     }
 }
