+++ src/main/java/com/powsybl/openloadflow/network/impl/LfNetworkLoaderImpl.java
@@ -405,8 +405,8 @@ public class LfNetworkLoaderImpl implements LfNetworkLoader<Network> {
                 ThreeWindingsTransformer.Leg leg = t3wt.getLeg(side);
                 LfBus lfBus = getLfBus(leg.getTerminal(), lfNetwork, parameters.isBreakers());
                 LfLegBranch lfBranch = LfLegBranch.create(lfNetwork, lfBus, lfBus0, t3wt, leg,
-                        topoConfig.retainPtc(LfLegBranch.getId(side, t3wt.getId())),
-                        topoConfig.retainRtc(LfLegBranch.getId(side, t3wt.getId())),
+                        topoConfig.isRetainedPtc(LfLegBranch.getId(side, t3wt.getId())),
+                        topoConfig.isRetainedRtc(LfLegBranch.getId(side, t3wt.getId())),
                         parameters);
                 addBranch(lfNetwork, lfBranch, report);
                 postProcessors.forEach(pp -> pp.onBranchAdded(t3wt, lfBranch));
